Ditch CoffeeScript and use Babel + Webpack
I don't consider CoffeeScript as a good step toward the future
as it would be nearly impossible to use some ES6 sweetness there.
For now the src file and test file has been transpiled to JS
as it is just to get rid of extra build step.

The lib folder is no longer in repository, it will be present in
NPM package only. This will contain only single file with Babel
transformed code to ensure ES5 compatibility.

For a browser the dist folder has been created. It contains UMD
bundle to be used anywhere without need to compile anything.

Removed paragraph from README about risks of using Symbol as with
Babel this is correctly handled now.